# api-docs.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "Uwazi API Documentation",
                 "version":  "1.0.0",
                 "description":  "API documentation for the Uwazi Social Justice Tracking Application"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:5000"
                    }
                ],
    "components":  {
                       "securitySchemes":  {
                                               "AdminAuth":  {
                                                                 "type":  "apiKey",
                                                                 "in":  "header",
                                                                 "name":  "X-User-Id",
                                                                 "description":  "Admin user ID used for protected admin endpoints"
                                                             }
                                           }
                   },
    "paths":  {
...
                  "/api/comments/report/{reportId}":  {
                                                          "get":  {
                                                                      "summary":  "Get all comments for a report",
                                                                      "tags":  [
                                                                                   "Comments"
                                                                               ],
                                                                      "parameters":  [
                                                                                         {
                                                                                             "in":  "path",
                                                                                             "name":  "reportId",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "List of comments for the report"
                                                                                                }
                                                                                    }
                                                                  }
                                                      },
...
                  "/api/reports":  {
                                       "post":  {
                                                    "summary":  "Create a new report",
                                                    "tags":  [
                                                                 "Reports"
                                                             ],
                                                    "requestBody":  {
                                                                        "required":  true,
                                                                        "content":  {
                                                                                        "application/json":  {
                                                                                                                 "schema":  {
                                                                                                                                "type":  "object",
                                                                                                                                "required":  [
                                                                                                                                                 "title",
                                                                                                                                                 "category",
                                                                                                                                                 "county"
                                                                                                                                             ],
                                                                                                                                "properties":  {
                                                                                                                                                   "title":  {
                                                                                                                                                                 "type":  "string"
                                                                                                                                                             },
                                                                                                                                                   "description":  {
                                                                                                                                                                       "type":  "string"
                                                                                                                                                                   },
                                                                                                                                                   "category":  {
                                                                                                                                                                    "type":  "string",
                                                                                                                                                                    "description":  "e.g. Judiciary, Police"
                                                                                                                                                                },
                                                                                                                                                   "county":  {
                                                                                                                                                                  "type":  "string",
                                                                                                                                                                  "description":  "Kenya county name"
                                                                                                                                                              },
                                                                                                                                                   "location":  {
                                                                                                                                                                    "type":  "string"
                                                                                                                                                                },
                                                                                                                                                   "isAnonymous":  {
                                                                                                                                                                       "type":  "boolean"
                                                                                                                                                                   },
                                                                                                                                                   "authorName":  {
                                                                                                                                                                      "type":  "string"
                                                                                                                                                                  },
                                                                                                                                                   "evidence":  {
                                                                                                                                                                    "type":  "array",
                                                                                                                                                                    "items":  {
                                                                                                                                                                                  "type":  "string",
                                                                                                                                                                                  "description":  "URL to photo or video evidence"
                                                                                                                                                                              }
                                                                                                                                                                }
                                                                                                                                               }
                                                                                                                            }
                                                                                                             }
                                                                                    }
                                                                    },
                                                    "responses":  {
                                                                      "201":  {
                                                                                  "description":  "Report created successfully"
                                                                              },
                                                                      "400":  {
                                                                                  "description":  "Validation error"
                                                                              }
                                                                  }
                                                },
                                       "get":  {
                                                   "summary":  "Get all reports",
                                                   "tags":  [
                                                                "Reports"
                                                            ],
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "List of reports"
                                                                             }
                                                                 }
                                               }
                                   },
                  "/api/reports/{id}":  {
                                            "get":  {
                                                        "summary":  "Get a single report by ID",
                                                        "tags":  [
                                                                     "Reports"
                                                                 ],
                                                        "parameters":  [
                                                                           {
                                                                               "in":  "path",
                                                                               "name":  "id",
                                                                               "required":  true,
                                                                               "schema":  {
                                                                                              "type":  "string"
                                                                                          }
                                                                           }
                                                                       ],
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Report details"
                                                                                  },
                                                                          "404":  {
                                                                                      "description":  "Report not found"
                                                                                  }
                                                                      }
                                                    },
                                            "put":  {
                                                        "summary":  "Update a report",
                                                        "tags":  [
                                                                     "Reports"
                                                                 ],
                                                        "parameters":  [
                                                                           {
                                                                               "in":  "path",
                                                                               "name":  "id",
                                                                               "required":  true,
                                                                               "schema":  {
                                                                                              "type":  "string"
                                                                                          }
                                                                           }
                                                                       ],
                                                        "requestBody":  {
                                                                            "required":  true,
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "type":  "object",
                                                                                                                                    "properties":  {
                                                                                                                                                       "title":  {
                                                                                                                                                                     "type":  "string"
                                                                                                                                                                 },
                                                                                                                                                       "description":  {
                                                                                                                                                                           "type":  "string"
                                                                                                                                                                       },
                                                                                                                                                       "category":  {
                                                                                                                                                                        "type":  "string"
                                                                                                                                                                    },
                                                                                                                                                       "county":  {
                                                                                                                                                                      "type":  "string"
                                                                                                                                                                  },
                                                                                                                                                       "location":  {
                                                                                                                                                                        "type":  "string"
                                                                                                                                                                    },
                                                                                                                                                       "status":  {
                                                                                                                                                                      "type":  "string"
                                                                                                                                                                  },
                                                                                                                                                       "priority":  {
                                                                                                                                                                        "type":  "string"
                                                                                                                                                                    },
                                                                                                                                                       "evidence":  {
                                                                                                                                                                        "type":  "array",
                                                                                                                                                                        "items":  {
                                                                                                                                                                                      "type":  "string"
                                                                                                                                                                                  }
                                                                                                                                                                    }
                                                                                                                                                   }
                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Report updated successfully"
                                                                                  },
                                                                          "404":  {
                                                                                      "description":  "Report not found"
                                                                                  }
                                                                      }
                                                    },
                                            "delete":  {
                                                           "summary":  "Delete a report",
                                                           "tags":  [
                                                                        "Reports"
                                                                    ],
                                                           "parameters":  [
                                                                              {
                                                                                  "in":  "path",
                                                                                  "name":  "id",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "Report deleted successfully"
                                                                                     },
                                                                             "404":  {
                                                                                         "description":  "Report not found"
                                                                                     }
                                                                         }
                                                       }
                                        },
...
              },
...
}